kernel: Fix kernel modules compilation with IPK
Kernel-modules are now suffixed with the kernel-version.
The kernel version is retrieved from kernel sources and it
uses EXTRAVERSION. In our case, it contains a "_" character
which is considered as "illegal". The build currently fails because
of this error.

Removing the kernel module suffix using KERNEL_MODULE_PACKAGE_SUFFIX
fixes the issue.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>